<SearchBarNext /> -> <SearchBar />, with deprecation warning.
